# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/auth/login":{"get":{"tags":["auth"],"summary":"Get Auth Login","operationId":"get_auth_login_auth_login_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/logout":{"get":{"tags":["auth"],"summary":"Get Auth Login","operationId":"get_auth_login_auth_logout_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/token":{"get":{"tags":["auth"],"summary":"Get Auth Token","operationId":"get_auth_token_auth_token_get","parameters":[{"name":"token","in":"query","required":false,"schema":{"type":"string","default":"","title":"Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/me":{"get":{"tags":["user"],"summary":"Get User","operationId":"get_user_user_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/user/me/":{"put":{"tags":["user"],"summary":"Get User","operationId":"get_user_user_me__put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Me"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/byId/:id":{"get":{"tags":["user"],"summary":"Get User By Id","operationId":"get_user_by_id_user_byId__id_get","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/post/":{"post":{"tags":["post"],"summary":"Get Newest Post","operationId":"get_newest_post_post__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/post/:id":{"put":{"tags":["post"],"summary":"Get Newest Post","operationId":"get_newest_post_post__id_put","parameters":[{"name":"post_id","in":"query","required":true,"schema":{"type":"integer","title":"Post Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PutPost"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/post/newest":{"get":{"tags":["post"],"summary":"Get Newest Post","operationId":"get_newest_post_post_newest_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/post/byId/:id":{"get":{"tags":["post"],"summary":"Get Post By Id","operationId":"get_post_by_id_post_byId__id_get","parameters":[{"name":"post_id","in":"query","required":true,"schema":{"type":"integer","title":"Post Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/post/followed":{"get":{"tags":["post"],"summary":"Get Post Of Followed","operationId":"get_post_of_followed_post_followed_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/follower/ofUser/:id":{"get":{"tags":["follower"],"summary":"Get Followers Of User","operationId":"get_followers_of_user_follower_ofUser__id_get","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/follower/follow/:id":{"post":{"tags":["follower"],"summary":"Get Followers Of User","operationId":"get_followers_of_user_follower_follow__id_post","parameters":[{"name":"follow_id","in":"query","required":true,"schema":{"type":"integer","title":"Follow Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/follower/unfollow/:id":{"post":{"tags":["follower"],"summary":"Get Followers Of User","operationId":"get_followers_of_user_follower_unfollow__id_post","parameters":[{"name":"follow_id","in":"query","required":true,"schema":{"type":"integer","title":"Follow Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/search/":{"get":{"tags":["search"],"summary":"Get Search","operationId":"get_search_search__get","parameters":[{"name":"aduser","in":"query","required":false,"schema":{"type":"string","title":"Aduser"}},{"name":"title","in":"query","required":false,"schema":{"type":"string","title":"Title"}},{"name":"teaser","in":"query","required":false,"schema":{"type":"string","title":"Teaser"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/comment/ofPost/:id":{"get":{"tags":["comments"],"summary":"Get Comments Of Post","operationId":"get_comments_of_post_comment_ofPost__id_get","parameters":[{"name":"post_id","in":"query","required":true,"schema":{"type":"integer","title":"Post Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/comment/post/:id":{"post":{"tags":["comments"],"summary":"Comment Post","operationId":"comment_post_comment_post__id_post","parameters":[{"name":"post_id","in":"query","required":true,"schema":{"type":"integer","title":"Post Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Comment"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/comment/comment/:id":{"post":{"tags":["comments"],"summary":"Comment Post","operationId":"comment_post_comment_comment__id_post","parameters":[{"name":"comment_id","in":"query","required":true,"schema":{"type":"integer","title":"Comment Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Comment"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/comment/:id":{"put":{"tags":["comments"],"summary":"Comment Post","operationId":"comment_post_comment__id_put","parameters":[{"name":"comment_id","in":"query","required":true,"schema":{"type":"integer","title":"Comment Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Comment"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Comment":{"properties":{"content":{"type":"string","title":"Content"}},"type":"object","required":["content"],"title":"Comment"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Me":{"properties":{"displayname":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Displayname"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","title":"Me"},"Post":{"properties":{"thumbnail":{"type":"string","title":"Thumbnail"},"teaser":{"type":"string","title":"Teaser"},"title":{"type":"string","title":"Title"},"content":{"type":"string","title":"Content"}},"type":"object","required":["thumbnail","teaser","title","content"],"title":"Post"},"PutPost":{"properties":{"thumbnail":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Thumbnail"},"teaser":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Teaser"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content"}},"type":"object","title":"PutPost"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}